Module traffloat_types::def::reaction

Reaction definitions

Structs

Catalyst

A condition or catalyst which affects the rate of a reaction.

Category

A category of reaction.

CategoryId

Identifies a reaction category

Multipliers

The multipliers associated with a catalyst.

Type

A type of reaction.

TypeId

Identifies a reaction category

Enums

CatalystRange

A type of resource whose existence affects a reaction.

Put

The inputs and outputs of a reaction.