Enum traffloat_types::def::reaction::CatalystRange

pub enum CatalystRange {
    Cargo {
        ty: TypeId,
        levels: Range<CargoSize>,
    },
    Liquid {
        ty: TypeId,
        levels: Range<LiquidVolume>,
    },
    Gas {
        ty: TypeId,
        levels: Range<GasVolume>,
    },
    Electricity {
        levels: Range<ElectricPower>,
    },
    Light {
        levels: Range<Brightness>,
    },
    Skill {
        ty: TypeId,
        levels: Range<Skill>,
    },
}

A type of resource whose existence affects a reaction.

Variants

Cargo

Existence of cargo

Fields of Cargo

ty: TypeId

Type of cargo catalyst

levels: Range<CargoSize>

Min and max levels of cargo catalyst

Liquid

Existence of liquid

Fields of Liquid

ty: TypeId

Type of liquid catalyst

levels: Range<LiquidVolume>

Min and max levels of liquid catalyst

Gas

Existence of gas

Fields of Gas

ty: TypeId

Type of gas catalyst

levels: Range<GasVolume>

Min and max levels of gas catalyst

Electricity

Existence of power

Fields of Electricity

levels: Range<ElectricPower>

Min and max levels of electricity catalyst

Light

Existence of light

Fields of Light

levels: Range<Brightness>

Min and max levels of light catalyst

Skill

Existence of skilled operators

Only the most skilled operator is counted as a catalyst.

Fields of Skill

ty: TypeId

Type of skill catalyst

levels: Range<Skill>

Min and max levels of skill catalyst